# backend/internal/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MT-Predictor Backend API",
    "description": "Хакатон Московского транспорта · Трек №3. Go Core: телеметрия NDTP, REST, WebSocket, DSS Holding.",
    "version": "0.2.0",
    "contact": {
      "name": "teamv39 / Денис"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Go backend"
    }
  ],
...
  "paths": {
...
    "/api/v1/recommendations/{id}/apply": {
      "post": {
        "tags": ["dss"],
        "summary": "Apply Holding recommendation",
        "operationId": "applyRecommendation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "example": "alert_001"
          }
        ],
        "responses": {
          "200": {
            "description": "Recommendation applied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "applied" },
                    "recommendation": { "type": "string" },
                    "holding_active": { "type": "boolean" },
                    "dispatched_to": { "type": "string" },
                    "timestamp": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/simulation/control": {
      "post": {
        "tags": ["simulation"],
        "summary": "Control demo feeder playback",
        "operationId": "simulationControl",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["action"],
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": ["play", "pause", "speed", "reset"]
                  },
                  "speed": {
                    "type": "number",
                    "example": 5.0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated"
          }
        }
      }
    },
...
  },
...
}